requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
gcloud==0.18.3
googleapis-common-protos==1.70.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.4.2
jwcrypto==1.5.6
kiwisolver==1.4.7
lightning==2.5.1.post0
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.4.3
networkx==3.2.1
numpy==2.0.2
oauth2client==4.1.3
onnx==1.18.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.0rc2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
torchmetrics==1.7.1
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.2
yarl==1.20.0
zipp==3.21.0
